linter fix

+++ contracts/Interfaces/Compound/CErc20I.sol
@@ -1,15 +1,25 @@
 pragma solidity >=0.5.0;
 
-import './CTokenI.sol';
+import "./CTokenI.sol";
 
-interface CErc20I is CTokenI{
-    function mint(uint mintAmount) external returns (uint);
-    function redeem(uint redeemTokens) external returns (uint);
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
-    function borrow(uint borrowAmount) external returns (uint);
-    function repayBorrow(uint repayAmount) external returns (uint);
-    function repayBorrowBehalf(address borrower, uint repayAmount) external returns (uint);
-    function liquidateBorrow(address borrower, uint repayAmount, CTokenI cTokenCollateral) external returns (uint);
+interface CErc20I is CTokenI {
+    function mint(uint256 mintAmount) external returns (uint256);
+
+    function redeem(uint256 redeemTokens) external returns (uint256);
+
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
+
+    function borrow(uint256 borrowAmount) external returns (uint256);
+
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
+
+    function repayBorrowBehalf(address borrower, uint256 repayAmount) external returns (uint256);
+
+    function liquidateBorrow(
+        address borrower,
+        uint256 repayAmount,
+        CTokenI cTokenCollateral
+    ) external returns (uint256);
 
     function underlying() external returns (address);
 }
